# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inventory Management API",
    "description": "API documentation for the Inventory Management System",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development server"
    },
    {
      "url": "https://inventory-app-zffs.onrender.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Item": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "The auto-generated id of the item"
          },
          "name": {
            "type": "string",
            "description": "The name of the item"
          },
          "description": {
            "type": "string",
            "description": "The description of the item"
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "The price of the item"
          },
          "quantity": {
            "type": "integer",
            "description": "The available quantity of the item"
          },
          "category": {
            "type": "string",
            "description": "The category of the item"
          },
          "image_url": {
            "type": "string",
            "description": "The URL of the item image"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "The creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "The last update timestamp"
          }
        },
        "required": ["name", "price", "quantity"]
      },
      "ItemInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Product Name",
            "description": "Name of the item"
          },
          "description": {
            "type": "string",
            "example": "Product description",
            "description": "Description of the item"
          },
          "price": {
            "type": "number",
            "format": "float",
            "example": 99.99,
            "description": "Price of the item"
          },
          "quantity": {
            "type": "integer",
            "example": 10,
            "description": "Available quantity"
          },
          "category": {
            "type": "string",
            "example": "Electronics",
            "description": "Category of the item"
          },
          "image": {
            "type": "string",
            "format": "binary",
            "description": "Image file of the item (jpg, jpeg, png, gif)"
          }
        },
        "required": ["name", "price", "quantity"]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/items": {
      "get": {
        "tags": ["Items"],
        "summary": "Get all items",
        "description": "Retrieves a list of all items in the inventory",
        "operationId": "getItems",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Items"],
        "summary": "Create a new item",
        "description": "Adds a new item to the inventory",
        "operationId": "createItem",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/items/{id}": {
      "put": {
        "tags": ["Items"],
        "summary": "Update an item",
        "description": "Updates an existing item in the inventory",
        "operationId": "updateItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the item to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Item updated"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Items"],
        "summary": "Delete an item",
        "description": "Removes an item from the inventory",
        "operationId": "deleteItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the item to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Item deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Item deleted"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}